Hosting metrics: move uplot distr to quadtree.js and fix types on multi bars uplot plugin
